Simplify UserListRow - Merge UserListRow to UserList.swift
